+++ Semana_4/node/src/modules/users/infra/http/routes/profile.routes.ts
@@ -1,0 +1,12 @@
+import { Router } from "express";
+import ProfileController from "@users/infra/http/controllers/ProfileController";
+import ensureAuthenticated from "@users/infra/http/middlewares/ensureAuthenticated";
+
+let profileRouter = Router();
+let profileController = new ProfileController();
+
+profileRouter.use(ensureAuthenticated);
+profileRouter.put("/update", profileController.create);
+profileRouter.get("/show", profileController.show);
+
+export default profileRouter;