+++ Semana_4/node/src/shared/infra/http/routes/index.ts
@@ -1,6 +1,7 @@
 import { Router, json } from "express";
 import appointmentsRouter from "@modules/appointments/infra/http/routes/appointments.routes";
 import usersRouter from "@users/infra/http/routes/users.routes";
+import profileRouter from "@users/infra/http/routes/profile.routes";
 import sessionsRouter from "@users/infra/http/routes/sessions.routes";
 import passwordRouter from "@users/infra/http/routes/password.routes";
 import LogRoutes from "@shared/infra/http/middlewares/logRoutes";
@@ -14,5 +15,6 @@ routes.use("/appointments", appointmentsRouter); // Repassa qualquer rota http q
 routes.use("/users", usersRouter); // Repassa qualquer rota http que ira passar pelo caminho /users para o router de users
 routes.use("/sessions", sessionsRouter);
 routes.use("/password", passwordRouter);
+routes.use('/profile', profileRouter)
 
 export default routes;